live + leaderboard
Starting work on implementing a leaderboard, so new view, controller
functions, etc. IT’s very incomplete, it won’t work as is. Also worth
noting, I am testing the project live at wold.comuf.com and it seems to
be going well so far. Database is working and the ui transferred over
pretty well. I will be actively updating both the site and the local
version.